Reimplemented PriorityQueue based on TreeSet<T> rather than Vector<T> to get automatic sorting of events.
Added dequeueUntil, isEmptyUntil etc. to handle the new criteria to only
process events up until a specific event (typically, the tick event). 

Expanded the Evaluation interface to include EventOccurrence. In all
classes, either implemented as the evaluate methods with an
EventOccurrence. In StateMachineSystem and in StateMachine, the old
evaluate() and evaluate(StateMachineSystem) now throws an exceptions.
This will be removed in the future.